Deprecate common test mixins
At the moment they make debugging more difficult
and do not enable embedding individual mock data into requests
