--- embed_video/tests/backends/__init__.py
@@ -1,16 +1,0 @@
-from embed_video.backends import detect_backend
-
-
-class BackendTestMixin:
-    urls = []
-    instance = None
-
-    def test_detect(self):
-        for url in self.urls:
-            backend = detect_backend(url[0])
-            self.assertIsInstance(backend, self.instance)
-
-    def test_code(self):
-        for url in self.urls:
-            backend = self.instance(url[0])
-            self.assertEqual(backend.code, url[1])
